식별값 풀 수정 & 테스트 코드 수정
